refactor : 애플리케이션들의 패키지를 분리
