+++ src/main/java/team/marco/voucher_management_system/console_app/ConsoleApplication.java
@@ -1,14 +1,15 @@
-package team.marco.voucher_management_system;
+package team.marco.voucher_management_system.console_app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.context.annotation.Import;
+import team.marco.voucher_management_system.VoucherManagementSystemApplication;
 import team.marco.voucher_management_system.console_app.application.CommandMainApplication;
 
+@Import(VoucherManagementSystemApplication.class)
 @SpringBootApplication
-@ConfigurationPropertiesScan
 public class ConsoleApplication {
     public static void main(String[] args) {
         SpringApplication springApplication = new SpringApplication(ConsoleApplication.class);
